requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6`